+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - feature: add support for configuring TLS settings for OTLP http exporter. (#26) | @codeboten
 - feature: instantiate a single SDK instead of one per signal. (#27) | @codeboten
+- feature: support `context` parameter in metric measurements. (#28) | @codeboten
 
 ## [0.1.10] - 2026-05-13